+++ src/main/java/com/github/theredbrain/healthregenerationoverhaul/mixin/entity/player/HungerManagerMixin.java
@@ -4,7 +4,6 @@ import com.github.theredbrain.healthregenerationoverhaul.HealthRegenerationOverh
 import net.minecraft.entity.player.HungerManager;
 import net.minecraft.entity.player.PlayerEntity;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Overwrite;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
@@ -17,7 +16,7 @@ public class HungerManagerMixin {
      */
     @Inject(method = "update", at = @At("HEAD"), cancellable = true)
     public void update(PlayerEntity player, CallbackInfo ci) {
-        if (HealthRegenerationOverhaul.serverConfig.disable_vanilla_food_system) {
+        if (HealthRegenerationOverhaul.SERVER_CONFIG.disable_vanilla_food_system) {
             ci.cancel();
         }
     }
@@ -27,7 +26,7 @@ public class HungerManagerMixin {
      */
     @Inject(method = "addExhaustion", at = @At("HEAD"), cancellable = true)
     public void addExhaustion(float exhaustion, CallbackInfo ci) {
-        if (HealthRegenerationOverhaul.serverConfig.disable_vanilla_food_system) {
+        if (HealthRegenerationOverhaul.SERVER_CONFIG.disable_vanilla_food_system) {
             ci.cancel();
         }
     }